\docType{data}
\name{probbase}
\alias{probbase}
\title{Conditional probability of InterVA4}
\format{A data frame with 246 observations on 81 variables.}
\description{
This is the table of conditional probabilities of symptoms
given CODs, together with the data check rules. The values are from InterVA-4.2.
}
\examples{
data(probbase)
}
\keyword{datasets}
